Rewrote configuration system
This previous system was good, but was difficult to change on the fly,
which is necessary for setting object defaults (see PR #329).
This system should support setting object defaults cleanly,
without needing separate classes for each type that will be configured.

The most relevant changes are in `nengo/config.py`. Now, instead of
subclassing, each backend should instantiate a `Config` object
and add extra parameters to it by calling `ExtraParams.set_param`.
`ExtraParams` is a class similar to the old `ConfigItem`, but it is
also designed to be extended live, instead of through subclassing.
It is also created automatically by the `Config` object;
`ExtraParams` shouldn't be instantiated manually.

Looking at the diff of `nengo/tests/test_config.py` should give
a good picture of how to use the new Config system.

Other changes:
- Networks in nef_summary example have unique labels now to avoid
  having the same hash and messing up Configs
- builder.Model keeps track of which network is the top level
  network, in order to do specific validation / optimization
- network.config is passed in to all build methods called
  by build_network (currently unused)
- Networks get a config object upon instantiation; bare networks
  inherit nengo.objects.toplevel_config, and subnetworks inherit
  the parent network's config
- Added a `params` decorator to NengoObject that returns a list
  of Parameters
- Made all appropriate NengoObject attributes `config.Parameter`s
  so that they share the implementation used by the config system
- Got rid of **kwargs in Connection.__init__
- Configs mess with pickling because of the WeakKeyDict.
  I think there are lots of reasons to not use Pickle though,
  so I got rid of saving / loading networks for now, and we should
  instead do something more intelligent in the future
- Fixed Uniform.__eq__ and Gaussian.__eq__ when comparing to
  other types of objects